move PersonName to its own file
