+++ lib/src/objects/contact_info.dart
@@ -1,5 +1,6 @@
 import 'package:mobile_scanner/src/objects/barcode.dart';
 import 'package:mobile_scanner/src/objects/email.dart';
+import 'package:mobile_scanner/src/objects/person_name.dart';
 
 /// A person's or organization's business card.
 /// For example a VCARD.
